backend_f/requirements.txt
Django==4.2.29
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
